+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - replaced the term "function" that to be differentiated with "differentiand", and changed related the parameter, functions, and arguments name. #89
 - replased the argument `same_subscripts_as_stencil` to `sort`. #74
 - updated the `equation` in `interpolation` module to sort the results. #74
+- replased the argument `evaluate` in the `equation` in `finite_difference` module with `keep_zero`. #80
+- changed the default value of `keep_zero` to `False`. #80
 
 ### Changes
 - replaced % format and str.format() with fstring. #71